+++ src/main/java/com/andersmmg/cc_modern/data/ModRecipeProvider.java
@@ -90,5 +90,29 @@ public class ModRecipeProvider extends RecipeProvider {
                 .requires(Items.GOLD_INGOT, 5)
                 .unlockedBy("has_computer_advanced", has(BuiltInRegistries.ITEM.get(COMPUTER_ADVANCED)))
                 .save(output, ResourceLocation.fromNamespaceAndPath(CCModern.MODID, "angled_monitor_advanced_from_normal"));
+
+        ShapedRecipeBuilder.shaped(RecipeCategory.REDSTONE, CCModern.TRANSPARENT_MONITOR_BLOCK.get(), 2)
+                .pattern("GGG")
+                .pattern("GSG")
+                .pattern("GGG")
+                .define('G', GLASS_PANES)
+                .define('S', Items.STONE)
+                .unlockedBy("has_computer_normal", has(BuiltInRegistries.ITEM.get(COMPUTER_NORMAL)))
+                .save(output, ResourceLocation.fromNamespaceAndPath(CCModern.MODID, "transparent_monitor"));
+
+        ShapedRecipeBuilder.shaped(RecipeCategory.REDSTONE, CCModern.TRANSPARENT_MONITOR_ADVANCED_BLOCK.get())
+                .pattern("GGG")
+                .pattern("GIG")
+                .pattern("GGG")
+                .define('G', GLASS_PANES)
+                .define('I', Items.GOLD_INGOT)
+                .unlockedBy("has_computer_advanced", has(BuiltInRegistries.ITEM.get(COMPUTER_ADVANCED)))
+                .save(output, ResourceLocation.fromNamespaceAndPath(CCModern.MODID, "transparent_monitor_advanced"));
+
+        ShapelessRecipeBuilder.shapeless(RecipeCategory.REDSTONE, CCModern.TRANSPARENT_MONITOR_ADVANCED_BLOCK.get())
+                .requires(CCModern.TRANSPARENT_MONITOR_BLOCK.get())
+                .requires(Items.GOLD_INGOT, 5)
+                .unlockedBy("has_computer_advanced", has(BuiltInRegistries.ITEM.get(COMPUTER_ADVANCED)))
+                .save(output, ResourceLocation.fromNamespaceAndPath(CCModern.MODID, "transparent_monitor_advanced_from_normal"));
     }
 }