+++ src/main/java/com/andersmmg/cc_modern/mixin/MonitorBlockEntityRendererAccess.java
@@ -2,14 +2,21 @@ package com.andersmmg.cc_modern.mixin;
 
 import dan200.computercraft.client.render.monitor.MonitorBlockEntityRenderer;
 import dan200.computercraft.client.render.monitor.MonitorRenderState;
+import dan200.computercraft.client.render.text.DirectFixedWidthFontRenderer;
+import dan200.computercraft.client.render.vbo.DirectVertexBuffer;
 import dan200.computercraft.core.terminal.Terminal;
 import dan200.computercraft.shared.peripheral.monitor.ClientMonitor;
 import org.joml.Matrix4f;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Invoker;
 
+import java.util.function.Consumer;
+
 @Mixin(MonitorBlockEntityRenderer.class)
 public interface MonitorBlockEntityRendererAccess {
     @Invoker("renderTerminal")
     void cc_modern$renderTerminal(Matrix4f matrix, ClientMonitor monitor, MonitorRenderState renderState, Terminal terminal, float xMargin, float yMargin);
+
+    @Invoker("renderToBuffer")
+    void cc_modern$renderToBuffer(DirectVertexBuffer vbo, int size, Consumer<DirectFixedWidthFontRenderer.QuadEmitter> draw);
 }